In progress

+++ celesta-sql/src/main/java/ru/curs/celesta/dbutils/adaptors/column/FireBirdColumnDefiners.java
@@ -85,7 +85,7 @@ class FireBirdBooleanColumnDefiner extends FireBirdColumnDefiner {
 
     @Override
     public String dbFieldType() {
-        return "numeric";
+        return "SMALLINT";
     }
 
     @Override
